requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
Pillow==10.1.0
huggingface-hub==0.19.4
flask-cors==4.0.0
gunicorn==21.2.0
